Moved function to script file + added tool for homolog
