+++ tools/admin_data/.gitignore
@@ -1,4 +1,4 @@
 *
 */
 !.gitignore
-
+!ALL_SOURCES_ALL_FREQUENCIES_genes_to_phenotype.txt
